aura-backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic-settings
python-dotenv
requests
openai
email-validator
argon2-cffi
PyJWT
